+++ src/main/java/blackjack/card/Deck.java
@@ -1,0 +1,4 @@
+package blackjack.card;
+
+public class Deck {
+}
