+++ src/main/java/blackjack/card/Symbol.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.card;
 
 public enum Symbol {
     ACE(1, "A"),
