+++ src/main/java/blackjack/user/Dealer.java
@@ -1,4 +1,6 @@
-package blackjack;
+package blackjack.user;
+
+import blackjack.card.Card;
 
 import java.util.ArrayList;
 import java.util.Collections;
