+++ src/main/java/blackjack/user/User.java
@@ -1,4 +1,6 @@
-package blackjack;
+package blackjack.user;
+
+import blackjack.card.Card;
 
 import java.util.List;
